Refactor export related functionality into separate class
