+++ mod-erm-usage-server/src/main/java/org/folio/rest/impl/CounterReportAPI.java
@@ -2,18 +2,17 @@ package org.folio.rest.impl;
 
 import static io.vertx.core.Future.succeededFuture;
 import static org.folio.rest.util.Constants.TABLE_NAME_COUNTER_REPORTS;
+import static org.folio.rest.util.ReportExportHelper.createDownloadResponseByReportVersion;
+import static org.folio.rest.util.ReportExportHelper.createExportMultipleMonthsResponseByReportVersion;
+import static org.folio.rest.util.ReportExportHelper.createExportResponseByFormat;
 import static org.folio.rest.util.VertxUtil.executeBlocking;
 
-import com.google.common.io.ByteStreams;
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Context;
 import io.vertx.core.Future;
 import io.vertx.core.Handler;
 import io.vertx.core.Promise;
-import io.vertx.core.json.Json;
 import io.vertx.sqlclient.Tuple;
-import java.io.IOException;
-import java.io.InputStream;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -22,7 +21,6 @@ import java.util.Comparator;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
@@ -43,25 +41,16 @@ import org.folio.rest.persist.Criteria.Criteria;
 import org.folio.rest.persist.Criteria.Criterion;
 import org.folio.rest.persist.PgUtil;
 import org.folio.rest.persist.cql.CQLWrapper;
-import org.folio.rest.tools.utils.BinaryOutStream;
 import org.folio.rest.tools.utils.ValidationHelper;
 import org.folio.rest.util.Constants;
 import org.folio.rest.util.PgHelper;
 import org.folio.rest.util.UploadHelper;
-import org.niso.schemas.counter.Report;
-import org.olf.erm.usage.counter.common.ExcelUtil;
-import org.olf.erm.usage.counter41.Counter4Utils;
-import org.olf.erm.usage.counter41.Counter4Utils.ReportMergeException;
-import org.olf.erm.usage.counter50.Counter5Utils;
-import org.olf.erm.usage.counter50.Counter5Utils.Counter5UtilsException;
 
 public class CounterReportAPI implements org.folio.rest.jaxrs.resource.CounterReports {
 
   private static final List<String> SUPPORTED_FORMATS = Arrays.asList("csv", "xlsx");
   private static final String UNSUPPORTED_MSG = "Requested format \"%s\" is not supported.";
-  private static final String UNSUPPORTED_COUNTER_VERSION_MSG =
-      "Requested counter version \"%s\" is not supported.";
-  private static final String XLSX_ERR_MSG = "An error occured while creating xlsx data: %s";
+
   private final Logger logger = LogManager.getLogger(CounterReportAPI.class);
 
   private final Comparator<CounterReportsPerYear> compareByYear =
@@ -197,23 +186,6 @@ public class CounterReportAPI implements org.folio.rest.jaxrs.resource.CounterRe
         asyncResultHandler);
   }
 
-  private Response createDownloadResponseByReportVersion(CounterReport report) {
-    if (report.getRelease().equals("4")) {
-      String xmlReport = Counter4Utils.toXML(Json.encode(report.getReport()));
-      return Optional.ofNullable(xmlReport)
-          .map(r -> GetCounterReportsDownloadByIdResponse.respond200WithApplicationXml(xmlReport))
-          .orElse(null);
-    } else if (report.getRelease().equals("5")) {
-      String jsonReport = Json.encode(report.getReport());
-      return Optional.ofNullable(jsonReport)
-          .map(r -> GetCounterReportsDownloadByIdResponse.respond200WithApplicationJson(jsonReport))
-          .orElse(null);
-    } else {
-      return GetCounterReportsDownloadByIdResponse.respond500WithTextPlain(
-          String.format("Unsupported report version '%s'", report.getRelease()));
-    }
-  }
-
   @Override
   public void getCounterReportsDownloadById(
       String id,
@@ -312,17 +284,6 @@ public class CounterReportAPI implements org.folio.rest.jaxrs.resource.CounterRe
     reportsYear.sort(compareByYear);
     result.setCounterReportsPerYear(reportsYear);
     return result;
-  }
-
-  private Optional<String> csvMapper(CounterReport cr) throws Counter5UtilsException {
-    if (cr.getRelease().equals("4") && cr.getReport() != null) {
-      return Optional.ofNullable(
-          Counter4Utils.toCSV(Counter4Utils.fromJSON(Json.encode(cr.getReport()))));
-    } else if (cr.getRelease().equals("5") && cr.getReport() != null) {
-      return Optional.ofNullable(
-          Counter5Utils.toCSV(Counter5Utils.fromJSON(Json.encode(cr.getReport()))));
-    }
-    return Optional.empty();
   }
 
   @Override
@@ -459,51 +420,6 @@ public class CounterReportAPI implements org.folio.rest.jaxrs.resource.CounterRe
             });
   }
 
-  private Response createExportResponseByFormat(CounterReport cr, String format) {
-    try {
-      return csvMapper(cr)
-          .map(
-              csvString -> {
-                if ("xlsx".equals(format)) {
-                  try {
-                    InputStream in = ExcelUtil.fromCSV(csvString);
-                    BinaryOutStream bos = new BinaryOutStream();
-                    bos.setData(ByteStreams.toByteArray(in));
-                    return GetCounterReportsExportByIdResponse
-                        .respond200WithApplicationVndOpenxmlformatsOfficedocumentSpreadsheetmlSheet(
-                            bos);
-                  } catch (IOException e) {
-                    return GetCounterReportsExportByIdResponse.respond500WithTextPlain(
-                        String.format(XLSX_ERR_MSG, e.getMessage()));
-                  }
-                }
-                return GetCounterReportsExportByIdResponse.respond200WithTextCsv(csvString);
-              })
-          .orElse(
-              GetCounterReportsExportByIdResponse.respond500WithTextPlain(
-                  "No report data or no mapper available"));
-    } catch (Counter5UtilsException e) {
-      return GetCounterReportsExportByIdResponse.respond500WithTextPlain(e.getMessage());
-    }
-  }
-
-  private Response createExportMultipleMonthsResponseByFormat(String csvString, String format) {
-    if ("xlsx".equals(format)) {
-      try {
-        InputStream in = ExcelUtil.fromCSV(csvString);
-        BinaryOutStream bos = new BinaryOutStream();
-        bos.setData(ByteStreams.toByteArray(in));
-        return GetCounterReportsExportProviderReportVersionFromToByIdAndNameAndAversionAndBeginAndEndResponse
-            .respond200WithApplicationVndOpenxmlformatsOfficedocumentSpreadsheetmlSheet(bos);
-      } catch (IOException e) {
-        return GetCounterReportsExportProviderReportVersionFromToByIdAndNameAndAversionAndBeginAndEndResponse
-            .respond500WithTextPlain(String.format(XLSX_ERR_MSG, e.getMessage()));
-      }
-    }
-    return GetCounterReportsExportProviderReportVersionFromToByIdAndNameAndAversionAndBeginAndEndResponse
-        .respond200WithTextCsv(csvString);
-  }
-
   @Override
   public void getCounterReportsExportById(
       String id,
@@ -539,25 +455,6 @@ public class CounterReportAPI implements org.folio.rest.jaxrs.resource.CounterRe
               GetCounterReportsExportByIdResponse.respond400WithTextPlain(
                   String.format(UNSUPPORTED_MSG, format))));
     }
-  }
-
-  private Response createExportMultipleMonthsResponseByReportVersion(
-      List<CounterReport> reportList, String format, String version) {
-    String csv;
-    try {
-      if (version.equals("4")) {
-        csv = counter4ReportsToCsv(reportList);
-      } else if (version.equals("5")) {
-        csv = counter5ReportsToCsv(reportList);
-      } else {
-        return GetCounterReportsExportProviderReportVersionFromToByIdAndNameAndAversionAndBeginAndEndResponse
-            .respond400WithTextPlain(String.format(UNSUPPORTED_COUNTER_VERSION_MSG, version));
-      }
-    } catch (Exception e) {
-      return GetCounterReportsExportProviderReportVersionFromToByIdAndNameAndAversionAndBeginAndEndResponse
-          .respond500WithTextPlain(e.getMessage());
-    }
-    return createExportMultipleMonthsResponseByFormat(csv, format);
   }
 
   @Override
@@ -651,34 +548,6 @@ public class CounterReportAPI implements org.folio.rest.jaxrs.resource.CounterRe
     return new CQLWrapper(criterion);
   }
 
-  private String counter4ReportsToCsv(List<CounterReport> reports) throws ReportMergeException {
-    List<Report> c4Reports =
-        reports.stream()
-            .map(cr -> Counter4Utils.fromJSON(Json.encode(cr.getReport())))
-            .filter(Objects::nonNull)
-            .collect(Collectors.toList());
-    Report merge = Counter4Utils.merge(c4Reports);
-    return Counter4Utils.toCSV(merge);
-  }
-
-  private String counter5ReportsToCsv(List<CounterReport> reports) throws Counter5UtilsException {
-    List<Object> c5Reports =
-        reports.stream()
-            .map(this::internalReportToCOP5Report)
-            .filter(Objects::nonNull)
-            .collect(Collectors.toList());
-    Object merge = Counter5Utils.merge(c5Reports);
-    return Counter5Utils.toCSV(merge);
-  }
-
-  private Object internalReportToCOP5Report(CounterReport report) {
-    try {
-      return Counter5Utils.fromJSON(Json.encode(report.getReport()));
-    } catch (Counter5UtilsException e) {
-      throw new CounterReportAPIRuntimeException(e);
-    }
-  }
-
   private Future<List<CounterReport>> decodeBase64Report(String encodedData, Context vertxContext) {
     Promise<List<CounterReport>> result = Promise.promise();
     executeBlocking(
@@ -698,13 +567,6 @@ public class CounterReportAPI implements org.folio.rest.jaxrs.resource.CounterRe
         .onSuccess(result::complete)
         .onFailure(result::fail);
     return result.future();
-  }
-
-  private static class CounterReportAPIRuntimeException extends RuntimeException {
-
-    public CounterReportAPIRuntimeException(Throwable cause) {
-      super(cause);
-    }
   }
 
   private static class ReportUploadException extends RuntimeException {
